convert SatipDevices to std::vector

+++ device.h
@@ -21,8 +21,8 @@ class cSatipDevice : public cDevice, public cSatipPidStatistics, public cSatipBu
 public:
   static bool Initialize(int DeviceCount);
   static void Shutdown(void);
-  static unsigned int Count(void);
-  static cSatipDevice *GetSatipDevice(int CardIndex);
+  static size_t Count(void);
+  static cSatipDevice* GetSatipDevice(int CardIndex);
   static cString GetSatipStatus(void);
 
   // private parts
